+++ ts/features/wallet/bancomatpay/component/BPayWalletPreview.tsx
@@ -7,7 +7,7 @@ import bPayImage from "../../../../../img/wallet/cards-icons/bPay.png";
 import { Body } from "../../../../components/core/typography/Body";
 import { IOStyles } from "../../../../components/core/variables/IOStyles";
 import { GlobalState } from "../../../../store/reducers/types";
-import { CardLogoPreview } from "../../component/CardLogoPreview";
+import { CardLogoPreview } from "../../component/card/CardLogoPreview";
 import { navigateToBPayDetailScreen } from "../../../../store/actions/navigation";
 import { BPayPaymentMethod } from "../../../../types/pagopa";
 import { useImageResize } from "../../onboarding/bancomat/screens/hooks/useImageResize";
