+++ ts/features/wallet/privative/component/card/BasePrivativeCard.tsx
@@ -15,7 +15,7 @@ import { Monospace } from "../../../../../components/core/typography/Monospace";
 import { IOColors } from "../../../../../components/core/variables/IOColors";
 import I18n from "../../../../../i18n";
 import { isImageURISource } from "../../../../../types/image";
-import BaseCardComponent from "../../../component/BaseCardComponent";
+import BaseCardComponent from "../../../component/card/BaseCardComponent";
 import { useImageResize } from "../../../onboarding/bancomat/screens/hooks/useImageResize";
 
 type Props = {
